requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.0.3
psycopg2-binary==2.9.3
SQLAlchemy==1.4.36
requests==2.28.1  # For Mailgun API if needed